+++ src/fft.hpp
@@ -29,31 +29,31 @@ class fft {
         Ipp64f *transform(const SrcType<Ipp64f> & src, boost::shared_ptr<FFTBuf<Ipp64f> > dst, int order, int scaling = 1, int pscaling = 12);
 
         /* Alloc */
-        static Ipp8u *alloc(Ipp8u *d, int length);
-        static Ipp16s *alloc(Ipp16s *d, int length);
-        static Ipp32s *alloc(Ipp32s *d, int length);
-        static Ipp32f *alloc(Ipp32f *d, int length);
-        static Ipp64f *alloc(Ipp64f *d, int length);
-        static Ipp16sc *alloc(Ipp16sc *d, int length);
-        static Ipp32sc *alloc(Ipp32sc *d, int length);
-        static Ipp32fc *alloc(Ipp32fc *d, int length);
-        static Ipp64fc *alloc(Ipp64fc *d, int length);
+        static Ipp8u *alloc(Ipp8u *d, long int length);
+        static Ipp16s *alloc(Ipp16s *d, long int length);
+        static Ipp32s *alloc(Ipp32s *d, long int length);
+        static Ipp32f *alloc(Ipp32f *d, long int length);
+        static Ipp64f *alloc(Ipp64f *d, long int length);
+        static Ipp16sc *alloc(Ipp16sc *d, long int length);
+        static Ipp32sc *alloc(Ipp32sc *d, long int length);
+        static Ipp32fc *alloc(Ipp32fc *d, long int length);
+        static Ipp64fc *alloc(Ipp64fc *d, long int length);
 
         /* Zero mem */
-        static void zero_mem(Ipp16s *d, int length);
-        static void zero_mem(Ipp32s *d, int length);
-        static void zero_mem(Ipp32f *d, int length);
-        static void zero_mem(Ipp64f *d, int length);
-        static void zero_mem(Ipp16sc *d, int length);
-        static void zero_mem(Ipp32sc *d, int length);
-        static void zero_mem(Ipp32fc *d, int length);
-        static void zero_mem(Ipp64fc *d, int length);
+        static void zero_mem(Ipp16s *d, long int length);
+        static void zero_mem(Ipp32s *d, long int length);
+        static void zero_mem(Ipp32f *d, long int length);
+        static void zero_mem(Ipp64f *d, long int length);
+        static void zero_mem(Ipp16sc *d, long int length);
+        static void zero_mem(Ipp32sc *d, long int length);
+        static void zero_mem(Ipp32fc *d, long int length);
+        static void zero_mem(Ipp64fc *d, long int length);
 
         /* Free */
         static void free(void *p) { ippsFree(p); }
 
         /* Other funcs */
-        static int order_to_length(int order);
+        static long int order_to_length(int order);
     private:
         int _bufsize;
         std::stack<Ipp8u *> _buffers;
